+++ CHANGELOG.md
@@ -19,6 +19,7 @@ of each other or without specifying a specific platform (using default platform 
 - Changed log level from info to warn if updating of the resource skipped because of absence in the deployment output
 - Fixed deploying a new bundle if latest deploy output is missing
 - Fixed a bug when the latest deployment section was updated when the `deploy` command ended with the ABORTED status
+- Added empty deployment package handling for lambda layer
 
 # [1.14.1] - 2024-10-08
 - Added support lambda layers with runtime DotNet